requirements.txt
Flask>=3.0.0
pytest>=8.0.0
fpdf2>=2.7.0
allure-pytest>=2.13.0
pytest-html>=4.0.0
